+++ typescript.js
@@ -5,7 +5,6 @@ const error = 2;
 module.exports = {
 	extends: [
 		'./base.js', //
-		'plugin:@typescript-eslint/eslint-recommended',
 		'plugin:@typescript-eslint/recommended',
 		'plugin:@typescript-eslint/recommended-requiring-type-checking',
 	],
@@ -22,10 +21,31 @@ module.exports = {
 		},
 	},
 	rules: {
-		'no-unused-vars': off,
+		// Off
+		'react/jsx-sort-props': 'off',
+		'no-unused-vars': 'off',
 		'@typescript-eslint/explicit-function-return-type': 'off',
-		'@typescript-eslint/camelcase': 'off',
-		'@typescript-eslint/prefer-regexp-exec': off,
+		// '@typescript-eslint/camelcase': 'off',
+		'@typescript-eslint/prefer-regexp-exec': 'off',
+		'@typescript-eslint/no-explicit-any': 'off',
+		'@typescript-eslint/max-classes-per-file': 'off',
+		'@typescript-eslint/max-line-length': 'off',
+		'@typescript-eslint/no-empty-interface': 'off',
+		'@typescript-eslint/no-empty': 'off',
+		'@typescript-eslint/no-var-requires': 'off',
+		'@typescript-eslint/object-literal-sort-keys': 'off',
+		'@typescript-eslint/no-unsafe-member-access': 'off',
+		'@typescript-eslint/no-use-before-define': 'off',
+		'@typescript-eslint/restrict-plus-operands': 'off',
+		'@typescript-eslint/no-unsafe-assignment': 'off',
+		'@typescript-eslint/typedef': [
+			off,
+			{
+				default: ['property-declaration'],
+			},
+		],
+
+		// Errors
 		'@typescript-eslint/no-unused-vars': [
 			error,
 			{
@@ -36,51 +56,70 @@ module.exports = {
 			},
 		],
 		'@typescript-eslint/ban-ts-comment': error, // use eslint-disable-next-line instead
-		'@typescript-eslint/naming-convention': [
-			'error',
-			{
-				selector: 'interface',
-				format: ['PascalCase'],
-				custom: {
-					regex: '^I[A-Z]',
-					match: true,
-				},
-			},
-		],
-		'@typescript-eslint/no-explicit-any': off,
 		'@typescript-eslint/array-type': [
 			error,
 			{
 				default: 'array-simple',
 			},
 		],
-		'@typescript-eslint/max-classes-per-file': off,
-		'@typescript-eslint/max-line-length': off,
 		'@typescript-eslint/member-ordering': error,
-		// '@typescript-eslint/no-duplicate-variable': [
-		// 	error,
-		// 	{
-		// 		default: ['check-parameters'],
-		// 	},
-		// ],
-		'@typescript-eslint/no-empty-interface': off,
-		'@typescript-eslint/no-empty': off,
-		// '@typescript-eslint/no-shadowed-variable': error,
-		// '@typescript-eslint/no-switch-case-fall-through': error,
-		'@typescript-eslint/no-var-requires': off,
-		'@typescript-eslint/object-literal-sort-keys': off,
-		// '@typescript-eslint/switch-default': error,
-		'@typescript-eslint/typedef': [
-			off,
-			{
-				default: ['property-declaration'],
-			},
-		],
-		// '@typescript-eslint/variable-name': [
-		// 	error,
-		// 	{
-		// 		default: ['ban-keywords', 'check-format', 'allow-leading-underscore', 'allow-pascal-case'],
-		// 	},
-		// ],
+
+		//  Warnings - will become errors in next major version
+		'@typescript-eslint/no-empty-function': ['warn', { allow: ['arrowFunctions'] }],
+
+		// Off - no sure what to do with these
+		'@typescript-eslint/unbound-method': 'off',
 	},
+	overrides: [
+		{
+			files: ['*.ts', '*.tsx'],
+			rules: {
+				// Off
+				'react/prop-types': 'off',
+
+				//  Warnings - will become errors in next major version
+				'@typescript-eslint/restrict-plus-operands': 'warn',
+				'@typescript-eslint/no-unsafe-member-access': 'warn',
+				'@typescript-eslint/naming-convention': [
+					'warn',
+					{
+						selector: 'default',
+						format: ['camelCase'],
+					},
+					{
+						selector: 'variable',
+						format: ['camelCase', 'UPPER_CASE', 'PascalCase', 'snake_case'],
+						leadingUnderscore: 'allow',
+					},
+					{
+						selector: 'function',
+						format: ['camelCase', 'PascalCase'],
+						leadingUnderscore: 'allow',
+					},
+					{
+						selector: 'property',
+						format: ['camelCase', 'snake_case', 'PascalCase', 'UPPER_CASE'],
+						leadingUnderscore: 'allow',
+					},
+					{
+						selector: 'parameter',
+						format: ['camelCase', 'snake_case', 'PascalCase'],
+						leadingUnderscore: 'allow',
+					},
+					{
+						selector: 'typeLike',
+						format: ['PascalCase'],
+					},
+					{
+						selector: 'interface',
+						format: ['PascalCase'],
+						custom: {
+							regex: '^I[A-Z]',
+							match: true,
+						},
+					},
+				],
+			},
+		},
+	],
 };